requirement.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.17
django-bootstrap4==0.0.7
django-crontab==0.7.1
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.3.14
openapi-codec==1.3.2
pycurl==7.43.0.2
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
soupsieve==1.7.1
uritemplate==3.0.0
urllib3==1.24.1
channels==0.11.0
